requirements.txt
asgiref==3.3.1
Django==3.1.3
django-decouple==2.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Markdown==3.3.3
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
